requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
bcrypt==3.2.0
cffi==1.14.6
click==8.0.1
colorama==0.4.4
cryptography==3.4.8
fastapi==0.68.1
greenlet==1.1.1
h11==0.12.0
iniconfig==1.1.1
mysql-connector-python==8.0.26
packaging==21.0
pluggy==1.0.0
protobuf==3.17.3
py==1.10.0
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.5
six==1.16.0
SQLAlchemy==1.4.23
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
hatesonar==0.0.7
better_profanity==0.7.0